Add reload command to xHorde
It's possible to reload the whole hordes when (for example) we must
connect to an other server. It means that the xcraft-core-horde condif
can handle multiple topologies for one app.

For example:
  "xcraft-core-horde": {
    "hordes": ["polypheme-thrall"],
    "topology": {
      "polypheme-thrall": {
        "host": "vc.cresus.ch",
        "commanderPort": 10000,
        "notifierPort": 20000
      },
      "polypheme-thrall#test": {
        "host": "test.cresus.ch",
        "commanderPort": 10000,
        "notifierPort": 20000
      }
    }
  }

By providing topology = "test", it will connect to the test.cresus.ch
server (and disconnect from the first server by the way).